backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-docx
python-multipart
jinja2
google-generativeai
pypdf
